+++ src/helper/parse-file-extensions.ts
@@ -1,0 +1,15 @@
+export const parseFileExtensions = (fileExtensions: string[] | null) => {
+  if (fileExtensions == null) {
+    return null;
+  }
+
+  const result = fileExtensions
+    .map((ext) => ext.trim())
+    .filter((ext) => ext !== '');
+
+  if (result.length === 0) {
+    return null;
+  }
+
+  return result;
+};