+++ CHANGELOG.md
@@ -1574,3 +1574,8 @@ All notable changes to this project will be documented in this file.
 
 - fixed a sporadic NullPointerException in the parser occurring during the auto-compile phase of the create-ingame process on Windows
 - improved reliability of the auto-compile script in the create-ingame process on Windows by implementing retry logic for temporary file detection. This addresses premature exits in the file copy operation that caused occasional desynchronization
+
+## [3.6.24] - 16.05.2025
+
+- added option to define additional file extensions considered during file extension auto-resolution - thanks for the suggestion to [@EntitySeaker](https://github.com/EntitySeaker)
+- added option to define the name of the main output file - thanks for the suggestion to [@EntitySeaker](https://github.com/EntitySeaker)